third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/worker-src/dedicated-none-expected.txt

CONSOLE ERROR: line 91: Refused to create a worker from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/ping.js' because it violates the following Content Security Policy directive: "worker-src 'none'".

CONSOLE ERROR: line 91: Refused to create a worker from 'blob:http://127.0.0.1:8000/c50ac71a-85bd-4385-87e7-3f615ab5e73c' because it violates the following Content Security Policy directive: "worker-src 'none'".

This is a testharness.js-based test.
PASS Same-origin dedicated worker blocked by host-source expression. 
PASS blob: dedicated worker blocked by 'blob:'. 
Harness: the test ran to completion.
